requirements.txt
pandas
xgboost
seaborn
numpy
spotipy
flask
Celery
redis
apache-airflow
mlflow
flask_socketio